feature: 書籍登録APIの実装

+++ api/server/user/internal/application/output/util.go
@@ -2,9 +2,9 @@ package output
 
 // ListQuery - 一覧検索クエリのレスポンス
 type ListQuery struct {
-	Limit  int64
-	Offset int64
-	Total  int64
+	Limit  int
+	Offset int
+	Total  int
 	Order  *QueryOrder
 }
